ref(KUI-1176): consolidate and make use of semesterUtils
